A few minor changes.
Added graph_lock functionality to let subscribed nodes know when the entire blueprint has been applied.
During the application process, the graph_lock is set to true. When the application process is complete, the graph_lock is set to false.
Made the render_edge_types be added in the setup phase rather than in each process_blueprint call.

Committing prior to adding support for actions